requirements.txt
undetected-chromedriver~=3.5.5
selenium~=4.27.1
PrettyColorPrinter~=0.45
python-dotenv~=1.0.1
beautifulsoup4~=4.12.3
PyMySQL~=1.1.1